+++ packages/components/elements/textarea/Textarea.tsx
@@ -4,6 +4,7 @@ import { cn } from "@util/index";
 
 import { Skeleton } from "@elements/skeleton";
 
+import { HelperText } from "../helperText";
 import { Label, LabelProps } from "../label";
 
 export interface TextareaProps
@@ -35,14 +36,14 @@ const Textarea = React.forwardRef<HTMLTextAreaElement, TextareaProps>(
       isLoading,
       ...props
     },
-    ref
+    ref,
   ) => {
     return (
       <div
         className={cn(
           "textarea-main hawa-relative hawa-flex hawa-h-full hawa-w-full hawa-flex-col",
           !forceHideHelperText && "hawa-gap-2",
-          className
+          className,
         )}
       >
         <div className="hawa-flex hawa-flex-row hawa-justify-between">
@@ -62,28 +63,17 @@ const Textarea = React.forwardRef<HTMLTextAreaElement, TextareaProps>(
           <Skeleton style={{ height: 40 }} />
         ) : (
           <textarea
-          {...textareaProps}
+            {...textareaProps}
             className={cn(
               "hawa-flex hawa-min-h-[40px] hawa-h-[40px]  hawa-w-full hawa-rounded-md hawa-border hawa-border-input hawa-bg-background hawa-px-3 hawa-py-2 hawa-text-sm hawa-ring-offset-background placeholder:hawa-text-gray-400 placeholder:hawa-text-muted-foreground focus-visible:hawa-outline-none focus-visible:hawa-ring-2 focus-visible:hawa-ring-ring focus-visible:hawa-ring-offset-0 disabled:hawa-cursor-not-allowed disabled:hawa-opacity-50",
-              classNames?.textarea
+              classNames?.textarea,
             )}
             ref={ref}
           />
         )}
         <div className="hawa-flex hawa-flex-row hawa-justify-between">
           {/* Regular helper text */}
-          {!forceHideHelperText && (
-            <p
-              className={cn(
-                "hawa-my-0 hawa-text-start hawa-text-xs hawa-text-helper-color hawa-transition-all",
-                props.helperText
-                  ? "hawa-h-4 hawa-opacity-100"
-                  : "hawa-h-0 hawa-opacity-0"
-              )}
-            >
-              {props.helperText}
-            </p>
-          )}
+          {!forceHideHelperText && <HelperText helperText={props.helperText} />}
 
           {/* Character Counter */}
           {showCount && countPosition === "bottom" && (
@@ -95,7 +85,7 @@ const Textarea = React.forwardRef<HTMLTextAreaElement, TextareaProps>(
         </div>
       </div>
     );
-  }
+  },
 );
 Textarea.displayName = "Textarea";