+++ packages/storybook/stories/ElementsStories/PinInput.stories.tsx
@@ -6,24 +6,18 @@ import { PinInput } from "@sikka/hawa/elements/pinInput";
 const meta = {
   title: "Elements/Inputs/Pin Input",
   component: PinInput,
-  parameters: {
-    backgrounds: {
-      default: "offwhite",
-      values: [{ name: "offwhite", value: "#ededed" }],
-    },
-    layout: "centered",
-  },
 } satisfies Meta<typeof PinInput>;
 
 export default meta;
 type Story = StoryObj<typeof PinInput>;
 
 export const Default: Story = {
+  name: "Pin Input",
   render: (args) => <PinInput {...args} />,
   parameters: { backgrounds: { default: "offwhite" } },
   args: {
-    getPins: () => console.log("getting pins"),
-    width: "normal",
-    digits: 6,
+    maxLength: 12,
+    pattern: "^[0-9a-zA-Z]*$",
+    separatorPosition: 50,
   },
 };